Now when the user presses a resident's button, the information is sent to the resident profile template and displayed.
Also, the FABs function has now been altered to open a create new resident form for the user to fill in. After the user has submitted the form, A new resident button will appear displaying the name from the form. When that new button is clicked it, like the other resident buttons, sends and displays the information given by the user in the resident template fragment.

These changes have been made to Unit 1 (FirstFragment) and in the future the same features will be added to Unit 2 (SecondFragment)